[#115] Refactor BatchResult & BatchReturningResult
