+++ core/src/main/java/io/github/zero88/jooqx/BatchReturningResult.java
@@ -1,6 +1,8 @@
 package io.github.zero88.jooqx;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 import org.jetbrains.annotations.NotNull;
 
@@ -11,6 +13,10 @@ import org.jetbrains.annotations.NotNull;
  * @since 1.0.0
  */
 public interface BatchReturningResult<R> extends BatchResult {
+
+    static <R> BatchReturningResult<R> create(int total, List<R> results) {
+        return new BatchReturningResultImpl<>(total, Optional.ofNullable(results).orElseGet(ArrayList::new));
+    }
 
     @NotNull
     List<R> getRecords();
